changes made to entities

+++ crowdsourced-local-business-api/src/main/java/com/fetchdear/crowd_api/repositories/BusinessRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BusinessRepository extends JpaRepository<Business, Integer> {
+public interface BusinessRepository extends JpaRepository<Business, Long> {
 
     public Business findByName(String name);
     public void deleteByName(String name);
